Add configurable performance parameters
executionTimeoutMinutes
Specify the maximum number of device minutes that your run can use by
setting a timeout on each device.
Default is 150 minutes

videoRecording
Record the video output for each device.
Default is 'on'

performanceMonitoring
Record performance metrics for each device and display those results
in a graph.
Default is 'on'